requirements.txt
# I recommend installing torch manually from 
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
Pillow
tqdm
seaborn
